requirements.txt
# # YOLOv5 requirements
# # Usage: pip install -r requirements.txt
# Pillow
# # opencv-python-headless
# opencv-python
# # os
# # shutil
# streamlit
# # Base ----------------------------------------
# matplotlib>=3.2.2
# numpy>=1.18.5
# opencv-python>=4.1.1
# Pillow>=7.1.2
# PyYAML>=5.3.1
# requests>=2.23.0
# scipy>=1.4.1
# torch>=1.7.0
# torchvision>=0.8.1
# tqdm>=4.64.0
# # protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012

# # Logging -------------------------------------
# tensorboard>=2.4.1
# # wandb
# # clearml

# # Plotting ------------------------------------
# pandas>=1.1.4
# seaborn>=0.11.0

# # Export --------------------------------------
# # coremltools>=5.2  # CoreML export
# # onnx>=1.9.0  # ONNX export
# # onnx-simplifier>=0.4.1  # ONNX simplifier
# # nvidia-pyindex  # TensorRT export
# # nvidia-tensorrt  # TensorRT export
# # scikit-learn==0.19.2  # CoreML quantization
# # tensorflow>=2.4.1  # TFLite export (or tensorflow-cpu, tensorflow-aarch64)
# # tensorflowjs>=3.9.0  # TF.js export
# # openvino-dev  # OpenVINO export

# # Extras --------------------------------------
# ipython  # interactive notebook
# psutil  # system utilization
# thop>=0.1.1  # FLOPs computation
# # albumentations>=1.0.3
# # pycocotools>=2.0  # COCO mAP
# # roboflow


absl-py==1.2.0
altair==4.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
cycler==0.11.0
Cython==0.29.30
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.9.1
fastjsonschema==2.16.1
fonttools==4.33.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.48.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.9.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.3
ninja==1.10.2.3
notebook==6.4.12
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.0
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
scikit-image==0.19.2
scipy==1.8.1
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.3.0
streamlit==1.11.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
thop==0.1.1.post2207130030
tifffile==2022.5.4
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.0
torch==1.11.0
torchvision==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.20.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.1
widgetsnbextension==3.6.1
yacs==0.1.8
zipp==3.8.1